+++ docs/whats-new.rst
@@ -4,7 +4,7 @@ What's New
 v0.1.5 (???)
 ------------
 
-Implemented vertical mode methods.
+Implemented vertical mode methods and modal KE and PE spectra diagnostics.
 
 Implemented multi-layer subclass.