+++ docs/whats-new.rst
@@ -1,7 +1,7 @@
 What's New
 ==========
 
-v0.1.5 (???)
+v0.2.0 (27 April 2016)
 ------------
 
 Added compatibility with python 3.